Add files via upload
Added working CombinationManager.cs script, partially powered by chatGPT, (I think you could tell by the code). At first I was iterating through a list of Items using an iterative loop, but chatGPT replaced this with a foreach loop. I also decided that it would be better to use a dictionary, and use the droppedItem GameObjects instead of items, as they actually have a transform position, so I can then calculate the distance from eachother and the center of the grouped items.